requirements.txt
# Meat Quality Monitoring System - Python Dependencies

# Core Streamlit framework
streamlit>=1.28.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization libraries
plotly>=5.17.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Additional utilities
altair>=5.0.0

# MQTT Client for ESP32 communication
paho-mqtt>=1.6.1

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Database (SQLite is built-in, but these are for convenience)
# No additional packages needed for SQLite

# Date/time utilities
python-dateutil>=2.8.2
